Moved MessageBlock
it *technically* stores text
