+++ gradle.properties
@@ -1,3 +1,3 @@
 org.gradle.daemon=true
 org.gradle.jvmargs=-Xms256m -Xmx1024m
-archash=64304225a74ee3e2c2bd97fd1f23dc4042ba29e2
+archash=6164484526b4e0a9dce4c59b11bef3448e363cee
